v94/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cryptography==41.0.7
pyjwt[crypto]==2.8.0
pyyaml==6.0.1
requests==2.31.0
python-multipart==0.0.6
